+++ lib/lightning/usage_tracking/configuration.ex
@@ -1,0 +1,14 @@
+defmodule Lightning.UsageTracking.Configuration do
+  @moduledoc """
+  Persistence of Usage Tracker submission configuration
+
+  """
+  use Ecto.Schema
+
+  @primary_key false
+  schema "usage_tracking_configurations" do
+    field :instance_id, Ecto.UUID, autogenerate: true
+
+    timestamps()
+  end
+end